Separated StandardLP from MergeMPS
